refactored modules to lib
