+++ backend/api/profile/create.ts
@@ -1,5 +1,5 @@
 import { VercelRequest, VercelResponse } from "@vercel/node";
-import { respond } from "../../modules/utils/respond.js";
+import { respond } from "../../lib/utils/respond.js";
 import { getLoggedInUser } from "../../middlewares/auth.js";
 import Identity, { SchemaFields } from "../../models/Identity.js";
 import { withmiddleware } from "../../middlewares/withMiddleware.js";
